Change AWS CLI profile to use common configuration
The credentials get written out to /etc/aws_config, and a script is
added to /etc/profile.d to set the location of that config, which means
that the location will be known to all users. All this is because the
postgres user needs to be able to use the AWS CLI to do database
backups.