Pass environment variable to exec
This isn't an ideal solution, but it's due to the fact that the Puppet
exec type doesn't preserve the environment like a login shell does, and
the script that does the restoration is relying on this variable being
set. The variable could also be exported in the script, but one
solution doesn't really seem any better than the other.